[Refactor] arena-info-table.cpp/h をarena-entry.cpp/h に変えた
