+++ src/market/arena-entry.h
