+++ src/market/arena.cpp
@@ -7,7 +7,7 @@
 #include "floor/floor-mode-changer.h"
 #include "io/input-key-acceptor.h"
 #include "main/sound-of-music.h"
-#include "market/arena-info-table.h"
+#include "market/arena-entry.h"
 #include "market/building-actions-table.h"
 #include "market/building-util.h"
 #include "monster-floor/place-monster-types.h"
