+++ src/system/player-type-definition.cpp
@@ -1,6 +1,6 @@
 #include "system/player-type-definition.h"
 #include "floor/geometry.h"
-#include "market/arena-info-table.h"
+#include "market/arena-entry.h"
 #include "system/angband-exceptions.h"
 #include "system/monster-race-info.h" // @todo 暫定、後で消す.
 #include "system/redrawing-flags-updater.h"
